galera: Add {$GALERA_NODECOUNT} macro to override the dflt 3 nodecount.
Now you need to set GALERA_NODECOUNT on all Galera nodes that have a
different cluster size. If it concerns many hosts, you may want to use a
custom Template with only that macro. (See Zabbix macro
inheritance/override rules.)

NOTE: Since it now checks not <3 but <>{$NODECOUNT} you'll get an alert
for excess nodes too. This cannot be helped. If we used <{$NODECOUNT} we
would forget to set the macro value properly when adding the node to
Zabbix.