diskio: Add free disk space monitoring for various mounts
Previously we relied on the built-in mount detection and space finding.
This had trouble with long mount points. Also this kept detecting
container image mount points, which is way too many.

Previously, we also used the ZFS template to detect all datasets. But
that was no good either (especially, again, because of all the
containerd mounts).

This config tries to keep a middle ground with a better selection of
mount points to scan. Also, while we do have disk size and disk in use
in bytes, we do not activate it because we don't need the extra items.